Use a different strategy for avoiding PIPE closure errors.
If the calling program pipes the output but closes the pipe before the program
finishes, this will cause errors writing to the pipe.  By default, this would
cause an ugly exception to be printed to stderr.

Before this commit, we addressed the pipe issue by catching EPIPE and closing
sys.stdout.  However, this interferes with --action=print on multiple files.

After this commit, we address the pipe issue by registering a SIGPIPE handler.