Fix parsing of 'with' statements in python2.7.
Add a hack to work around python2.7 changing the definition of col_offset for
'with' AST nodes.

Consider the code:
   with X: pass
   ^0   ^5
In python2.6, col_offset is 0.
In python2.7, col_offset is 5.

py26 >>> ast.parse("with foo: pass").body[0].col_offset
0

py27 >>> ast.parse("with foo: pass").body[0].col_offset
5

This change was presumably made because python2.7 allows for multiple 'with'
clauses:

   with X, Y: pass

Since 'Y's col_offset isn't the beginning of the line, the authors of Python
presumably changed 'X's col_offset to also not be the beginning of the
line. If they had made the With ast node support multiple clauses, they
wouldn't have needed to do that, but then that would introduce an API change
in the AST. So it's understandable that they did that.

Since we use startpos for breaking lines, we need to set startpos to
the beginning of the line.