Fix print syntax in test_imports2s.py for Python 3
The prints aren't actually the same in Python 2, because there is no
__future__.print_function import, so they will print a tuple if run. However,
a __future__ import would change the test, so I left it alone. The print is
not actually executed. The tests in this file still don't completely work in
Python 3, but I don't think the print is actually executed.