Stop using hanging indent for long (40+ char) symbols, by default
Before this commit, an import statement with really long symbols was formatted
with a "hanging indent" in order to avoid exceeding 79 characters.
For example:
    from   foo.bar1                 import (x23456789a, x23456789a1, x23456789a12,
                                            x23456789a123)
    from   foo.bar2                 import (
        x23456789a123456789b123456789c123456789, z1, z2)

After this commit, by default, the above example is formatted like this:
    from   foo.bar1                 import (x23456789a, x23456789a1, x23456789a12,
                                            x23456789a123)
    from   foo                      import (x23456789a123456789b123456789c123456789,
                                            z1, z2)

There is now a new command-line option --hanging-indent:
  --hanging-indent=auto:   symbols >= 40 chars long trigger hanging indent;
                           the old behavior
  --hanging-indent=never:  never use hanging indent;
                           the new default behavior
  --hanging-indent=always: always use hanging indent

Subjective rationale for changing the default to --hanging-indent=never:
Strict adherence to 79-column rule was detrimental.  Turning off the
hanging-indent feature is a better default.  The hanging-indent feature was
hurting readability and causing code churn when imports are touched.  People
reading code that only see the first 80 characters of the line will still see
sufficient amount of the information.  Even with hanging indent, if a symbol
was *really* long, it would still exceed the max line length; since at some
point the 79-column rule has to be broken anyway, we can live with increasing
how often that happens, in return for better overall readability.

Suggested-by & Reviewed-by: taylor@deshaw.com, srp@deshaw.com