Fix most of the tests in test_py.py to work with Python 3
There are still a few tests which I believe will require fixes in _py.py
itself in order to work properly in Python 3.

I have added --output=silent to tests that use sys.stdout.write or
open(...).write. This is because in Python 2 these functions return None, but
in Python 3 they return the number of characters written, resulting in an
extra number in the output without this flag.

I have changed an "exec" test to use "if 1: print(...)" so that it is a
non-expression even in Python 3.

Some tests' output are a bit odd now, but I believe the behavior is expected.
For instance, 'py print 3 + 5' prints "3 + 5" in Python 3 instead of "8". This
is because print is a built-in function and the remaining arguments are passed
to it as arguments (with '+' passed as a string, i.e., print(3, '+', 5)). I
believe this is the expected behavior.

'py print "3 + 5"' produces print(3 + 5) as expected, but the 3 + 5 is
automatically evaluated in the output, so it appears as

[PYFLYBY] print(8)
8

This also appears to be expected behavior, since the default eval behavior is
"auto". However, I haven't read that code particularly closely yet, so this
may not actually be the desired output. Some of the remaining test failures
are similar in nature, so this may end up changing in a future commit.