Fix 'py print sys' test output for Python 3
The output is a bit odd, because in Python 3 print is a function, and the
autocall behavior uses the repr() of the object in the [PYFLYBY] line instead
of the input text. However, this behavior is already existing for other
functions (the only difference here is that print is a function in Python 3),
so I am testing it as is for now.