Support IPython 6.0+ tab completions
IPython 6.0+ (with some exceptions for minor versions) uses Jedi by default,
and has the name completers that pyflyby hooks disabled. Hooking into Jedi is
more difficult, since it works on the full Python code, so instead we just
re-enable the global and attr completers when using pyflyby, so that they can
still be hooked.