Use the config file to disable come prompt-toolkit features
This lets the features be disabled for the py tests as well, because py
doesn't recognize the specific IPython command line flags to disable them.
Unfortunately, it doesn't work with the Jupyter console tests, because it uses
a different config file, and doesn't support all the config options that the
regular IPython does.