Handle comms when %load_ext is used
The frontend(lab) attempts to open comm targets as
soon as it is loaded. If the user prefers to not enable pyflyby
via config file and does it in the middle of an ipython
session (using %load_ext pyflyby), the frontend doesn't know
that it should now attempt opening the comm target again.

Added a new custom comm message to initialize comms - this will
be opened by pyflyby, not the frontend, and it will tell the
frontend to re-initialize the comm handlers/targets.