Fix a test failure
The output of the latest IPython prints the text printed by the completer an
extra time, I think because of the way prompt-toolkit handles things printed
to stdout when the prompt is active. The text is already printed more times
than it should be, so one extra time isn't a big deal for now.