Unsplit a string in test_interactive.py
Currently tidy-imports cannot handle strings that are split in this way.
Fixing it to work will be quite complicated, so in order to prevent it from
crashing on the test code, we remove it here. The issue is the code that tries
to find the first line of split multiline strings, which doesn't handle
strings inside of larger expressions. That entire code will become unneeded
with Python 3.8, which uses the correct line number for multiline strings.

See issue #12.