Fix tidy-imports with non-ASCII strings in Python 2
See #51 and https://bugs.python.org/issue18870. In order to parse a Unicode
string correctly, it has to be prepended with a PEP 263 encoding comment. It
would also work to not decode the string first, except that fails when there
is already an encoding argument, which there will be for source files that
have Unicode characters. Otherwise, Python 2 mixes up the encoded and decoded
versions of the Unicode characters, and the multiline string detecting code
fails.