Make sure to handle out of order call syntax
Also add test.

Call argument order are lost by AST; reorder them.

We also skip this test on earlier version of Python where this syntax is
anyway invalid; and pin promtp_toolkit to <3.0.6 – which we'll
investigate later.

We also make sure to print the pytest 10 slowest test to make sure to
keep an eye on performance.