Deprecate and stop using construction CompilerFlags from int.
Flags values have changed in Python in 3.9 and (backported to) 3.8.3,
thus constructing CompilerFlags from literal int is wrong.

As we can't detect literal we need to deprecate constructing from int.

Why do we not hardcode the all values of compiler flags from pre-3.9 ?

Because unfortunately CompilerFlags is a subclass of int which get
passed to compile(), function so we have to store the correct values,
and unfortunately they are also used as constructor for themselves
So we really can't distinguish values that would be correct from
incorrect.

This is made worse by the fact that the repr itself contains the hex
values, which really depends on the Python version and is thus
uiquitous.