Fix simple case of list comprehension name def
This change the resolving of names from deferring to actually try to be
greedy, and fallback to lazy if cant' resolve.

This fixes case like

    $ tidy-imports <<EOF
    from __future__ import absolute_import, division, print_function
    def f():
        x = 1
        y = [x for _ in range(5)]
        del x
    EOF
    [PYFLYBY] /dev/stdin:4: undefined name 'x' and no known import for it
    from __future__ import absolute_import, division, print_function
    def f():
        x = 1
        y = [x for _ in range(5)]
        del x

but will likely fail in other case like

    def f():
        def g():
            y = [x for _ in range(5)]
            return y
        x = 1
        g()
        del x

This will also change the behavior of:

    def F1():
         x1 = x2 = None
         def F2():
             x1, x2, x3, y1
         del x2

Where x2 is now going to be resolved instead of not. This is ambiguous
as we don't know when F2 will be called and thus wether x2 needs to be
imported or not.