Don't mute Ctrl-C in debugger.
This stop ignoring Ctrl-C when setting stdin to raw,
and allow us to generate a KeyboardInterrupt.

This will not propagate Ctrl-C to the underlying terminal anymore,
so might degrade the experience of interacting with GDB.

I did not manage to get a working instance of the remote debbuger
it segfault immediately for me. But this seem to adress the issue
when the process get stuck when trying to attach.