Make sure that the tty is correctly reset at the end of debugging.
This fixes a problem which happens when you run `py -d <PID>` to
attach a debugger to an external process: when the debugger
exits, the terminal is left in an unusable state and you need
either `stty sane` or `reset` to go back to normality.

The problem originates from the fact that Pty.communicate puts
the tty into "raw mode": in theory, there is code to restore the
tty to the previous state, but in practice it is not enough
because the Popen()ed gdb ALSO messes up with the tty
settings. This is a resonably accurate description of what
happened before this commit:

1. attach_debugger() Popen() a gdb process to inject some code into the
   debuggee process and ultimately call pdb.set_trace(). The Popen() does NOT
   redirect stdin, so gdb thinks it is attached to the terminal.

2. in the debugger process, Pty.communicate saves the current state of the tty
   and puts it into raw mode

3. the user quits the debugger.

4. gdb tries to display its prompt, but before doing that saves the state of the tty
   (which is in raw mode)

5A. Pty.communicate() ends, and reset the tty to the original state (saved in point 2)

5B. gdb is about to exit, and before doing so it resets the tty to what it
     thinks it was its original state saved in point 4. But at that point the
     tty was in raw mode, so gdb puts it in raw mode again.

Points 5A and 5B can happen in either order: whoever does it last, "wins". I
could easily force 5A to happen last by putting a time.sleep just before the
call to tcsetattr().

The correct fix is to redirect gdb stdin as well: this way gdb is completely
detached from the user tty and never tries to mess up with it.

A related change is to use --interpreter=mi. It has no effects at this point
because we never try to parse gdb output, but it is "the right thing to do"
when trying to control gdb from a pipe.