Do not access f_locals, which reset changes.
Given a file

    $ cat simple.py
    import ipdb

    def simple_f():
        myvar = 1
        print(myvar)
        ipdb.set_trace()
        print(myvar)

    if __name__ == '__main__':
        simple_f()

And running IPython with the pyflyby extension, we observe that changes
of values are not propagates to current frame:

    In [1]: %run -d simple.py
    NOTE: Enter 'c' at the ipdb>  prompt to continue execution.
    > simple.py(1)<module>()
    ----> 1 import ipdb
          2
          3 def simple_f():
          4     myvar = 1
          5     print(myvar)

    ipdb> c
    1
    > simple.py(7)simple_f()
          5     print(myvar)
          6     ipdb.set_trace()
    ----> 7     print(myvar)
          8
          9

    ipdb> myvar
    1
    ipdb> myvar = 2
    ipdb> myvar
    1
    ipdb> c
    1

this is due to the fact that accessing `curframe.f_locals` has side
effects and reset the value to the original one. CPython seem to use
`self.curframe_locals` to this effect to avoid side effects.

I'm not too sure of the effect this will have on autoreload. It is
also not the only things that access f_locals, in Particular IPython it
self does it in a couple of place to for example hide frames, and thus
currently ipdb's `where` also reset the default values. Thereforre
IPython will need to be modified as well.