Fix #147 – ClassDef not seen in current class.
This makes the class Definition visible inside its own methods,
it is likely to be a bit too broad as the class name might be visible in
its own body which is incorrect, but a rarer use case than referencing
cls in methods

that is to say the following will not try to import decimal.

    class Decimal:
        def foo(self):
            Decimal.x = 1

But now the following will not insert bar import for you:

    class Bar:

        @Bar
        def method(self):
            pass

There will be a few other similar cases, in particular because the rules
are not the same for top level classes and nested classes, and we are
leaning toward more false negative than false positive as users can
still insert the import themselves if missing, but autoimport might have
unattended side effects.