Allow debugger to not kill debugging process (#208)
* Allow debugger to not kill debugging process

wait_for_debugger_to_attach leaves a process behind for attaching a
debugger to it that stops as soon as the debugger is attached. To
facilitate attaching more debugger after initial debugging is done,
we now prompt at the end of debugging to check whether to continue
waiting for the debugger to be attached and do as per the user response.