Revert 6ee02b8af32e9c2a23ca286024d88380c3ba0403 commit (#218)
* Revert 6ee02b8af32e9c2a23ca286024d88380c3ba0403 partly

While attaching a debugger to already running process using '-d' option,
we exit the GDB launched using inject and the debugging happens outside
of the GDB.

So the GDB is exited but becomes a zombie process. But we have checks to exit
the debugging if the GDB has exited. Our checks to make sure if a process is
running or not doesn't treat zombie process as dead and hence it was a silent
pass through earlier.

With commit 6ee02b8af32e9c2a23ca286024d88380c3ba0403 the issue got exposed.
So temporarily reverting the change to prevent hampering any existing use cases.