Test on more recent Python versions.
Test on more recent Python versions.

This does a number of things:

  - raise pytest pinning to version 8,
  - test CI on 3.11
  - update deprecated formatargspec to use signature
  - update various ast node to use Constant, and .value instead of .s

We also mark a few test as xfail – as they have been failing for a
while, but IPython was indirectly pinned via prompt_toolkit).

And we also start to add a couple of compatibility to bytecode for
Python 3.12.