Don't execute test that will be skipped.
Having test that are skipped in not the best as it makes it hard to
track down if they might be ran on some platform combinaison.

Modify the fixture to just not run those test.

This make the test suite show less `s` for skipped tests, and make it
easier to see discrepancies.