Fix scope count for nested classes.
This was triggering negative number of class scopes,
and not allowing proper scoping of type annotations in class
definitions.

It is still weird that this work before unhiding the class scope, but I
guess as we are in Function Scope this is going to be a delayed
evaluation of the scope.