Refactor in view of import sorting
The parsing of statements is incorrect (comments statements soemtime
start with a new line).

While trying to investigate why, try to simplify/annotate the code a
bit, and type annotate it to detect inconsistency.

Also remove "optimisations" that assign attributes in top of cached
property and make reasoning about the code a bit too hard.