Misc cleanup and type simplification.
Trying to always get the same types in the same places,
this makes it a bit easier to reason bout the code.
It will let us also use static type checkers to verify modifications.

In the end we should likely remove polymorphism to avoid
redundant checks, or condition to transform object types.