Do not generate comment nodes that star with newlines.
This means that in some case, where we used to return a single
statement, we now may return multiple ones.

This also mean that a number of tests that were relying on identity now
need to rely on equality.

I'm not sure in the first place when you would rely on identity.