Better handling of unsafe filenames
See #346,

The handling of unsafe Filenames is a bit weird as the safety is checked
a bit too late in my opinion, and checked before the path is fully
resolved.

This tries to change this a bit by eing stricter on checking for
safeness, and refactor the caching of import db to use Pathlib (which
did not exists when FIlename was crreated).

This is still a bit more complicated than it should and might need some
cleanup later on.

There is also a buch of "deprecated" code in a if True in this area,
that would need propoer deprecation.