[New] saveframe utility
This commit adds a new utility named 'saveframe' in pyflyby.

This utility can be used to save information for debugging / reproducing
an issue.

Usage:
If you have a piece of code that is currently failing due to an issue originating
from another team's code, and you cannot share your private code as a reproducer,
use this utility to save relevant information to a file. Share the generated file
with the other team, enabling them to reproduce and diagnose the issue
independently.

This utility is provided with 2 interfaces: a script and a function.

Script location: `pyflyby/bin/saveframe`
Function location: `pyflyby.saveframe` (`from pyflyby import saveframe`)

For more information on the `savefrane` script, checkout the help message
using `pyflyby/bin/saveframe --help`

For more information on the `saveframe` function, checkout the doc of
`pyflyby.saveframe`.