Try to detect non-gloabl unused imports.
We do so by tracking the imports on exist of each scopestack; and on
exist, if there is an import that has not been checked; we mark it as
unused.

This seem to have side effects as the SourceToSource transformer is
aware it cannot remove non-global imports and emit a error message.

Note that this also affects conditional imports:

```
if sys.version_info > X,y:
    from foo import bar
else:
    from foo.new import bar
```

The first import will be marked as unused as shoadowed by the second one
(pyflyby does not understand the if/else).

This also does some refactor and type anotations while I am at it.