Fast module iter: use try/execept (#424)
This uses a try/except approach instead of a permission checking
to catch various problems that could arise preventing reading
dirs. It also adds a test.

I'm not sure how to properly test in a reliable
way when the process group is not the owner, at least on CI.

Locally:

$ pip install numpy
$ mkdir -p ./tmp/resricted
$ chmod 700 ./tmp/restricted
$ sudo -u nobody PYFLYBY_DISABLE_CACHE="1" PYFLYBY_SUPPRESS_CACHE_REBUILD_LOGS="1" PYTHONPATH="./tmp/restricted:$PYTHONPATH" py
In: [1]: import num<tab>

And check the importer is not disabled.